requirements.txt
alembic==1.11.1
atomicwrites==1.4.1
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
certifi==2024.8.30
cffi==1.16.0
charset-normalizer==3.4.0
click==8.1.3
colorama==0.4.6
coverage==7.2.7
cryptography==41.0.1
dnspython==2.3.0
dominate==2.8.0
email-validator==2.0.0.post2
exceptiongroup==1.2.2
Flask==2.3.2
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-Mail==0.10.0
Flask-Mailman==1.1.1
Flask-Migrate==4.0.4
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
freesms==0.2.0
greenlet==3.0.3
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
more-itertools==9.1.0
packaging==23.1
passlib==1.7.4
Pillow==10.0.0
pluggy==1.2.0
py==1.11.0
pycparser==2.21
pyparsing==3.1.0
pytest==7.4.0
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-editor==1.0.4
requests==2.32.3
scrypt==0.8.24
six==1.16.0
SQLAlchemy==2.0.17
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.3
visitor==0.1.3
wcwidth==0.2.6
Werkzeug==2.3.6
WTForms==3.0.1
